rename(user): 파일 위치 lib->components
