+++ apps/user/src/app/layout.tsx
@@ -2,7 +2,7 @@ import Provider from '@/components/Provider';
 import GoogleAnalytics from '@/lib/GoogleAnalytics';
 import StyledComponentRegistry from '@/lib/registry';
 import QueryClientProvider from '@/services/QueryClientProvider';
-import MobileAccessController from '@/lib/MobileAccessController';
+import MobileAccessController from '@/components/common/MobileAccessController/MobileAccessController';
 import React from 'react';
 import type { ReactNode } from 'react';
